// app/generated/RehabRequest.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: RehabRequest
// ====================================================

export interface RehabRequest_rehabRequest {
  __typename: "RehabRequest";
  id: string;
  images: (string | null)[] | null;
}

export interface RehabRequest {
  rehabRequest: RehabRequest_rehabRequest;
}

export interface RehabRequestVariables {
  query: string;
}